src/requirements.txt
numpy==1.22.3
pandas==1.4.2
pathlib==1.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
regex==2022.4.24
six==1.16.0
slack-bolt==1.13.2
slack-sdk==3.16.1
thefuzz==0.19.0
